+++ db-dumps/restore.sh
@@ -4,5 +4,5 @@ if [ -z "$1" ]; then
     exit 1
 fi
 DB=withings
-CMD="mysql -u user -p -h localhost --port=3306"
+CMD="mysql -u user -p -h localhost --protocol=tcp --port=3306"
 ${CMD} withings < $1